+++ data-raw/logoText.R
@@ -1,10 +1,10 @@
 
 logoTextMD <- "{{< fa head-side-virus >}}**&nbsp;&nbsp;I&nbsp;D&nbsp;E&nbsp;A&nbsp;S**<br>
-[**I**nfectious **D**isease **E**pidemiology and **A**pplied **S**tatistics]{.serif .smaller}"
+[**I**nfectious **D**isease **E**pidemiology & **A**pplied **S**tatistics]{.serif .smaller}"
 
 usethis::use_data(logoTextMD, overwrite = TRUE)
 
 
-logoTextHTML <- '<span class="sans-serif" style="font-size: 1em;"><i class="fa-solid fa-head-side-virus" aria-label="head-side-virus"></i><span style="font-weight: 700;">&nbsp;&nbsp;I&nbsp;D&nbsp;E&nbsp;A&nbsp;S</span></span> <br> <span class="serif" style="font-size: 0.9em;"><span style = "font-weight: 700;">I</span>nfectious <span style = "font-weight: 700;">D</span>isease <span style = "font-weight: 700;">E</span>pidemiology and <span style = "font-weight: 700;">A</span>pplied <span style = "font-weight: 700;">S</span>tatistics</span>'
+logoTextHTML <- '<span class="sans-serif" style="font-size: 1em;"><i class="fa-solid fa-head-side-virus" aria-label="head-side-virus"></i><span style="font-weight: 700;">&nbsp;&nbsp;I&nbsp;D&nbsp;E&nbsp;A&nbsp;S</span></span> <br> <span class="serif" style="font-size: 0.9em;"><span style = "font-weight: 700;">I</span>nfectious <span style = "font-weight: 700;">D</span>isease <span style = "font-weight: 700;">E</span>pidemiology & <span style = "font-weight: 700;">A</span>pplied <span style = "font-weight: 700;">S</span>tatistics</span>'
 
 usethis::use_data(logoTextHTML, overwrite = TRUE)